+++ src/js/src/main.js
@@ -22,7 +22,7 @@ import {
     faPlayCircle, faStopCircle, faPause, faDatabase, faSave, faCamera, faHistory, faSkullCrossbones, faUndoAlt, 
     faSyncAlt, faPowerOff, faPencil, faArrowRight, faArrowLeft, faCompactDisc, faCheckCircle, faHdd, faMinus, faTerminal,
     faPaintbrush, faTv, faCircle, faRefresh, faCaretDown, faTimesCircle, faAngleLeft, faAngleRight, faCopy,
-    faTableColumns, faArrowPointer, faBroom, faEraser
+    faTableColumns, faArrowPointer, faBroom, faEraser, faBook
 } from '@fortawesome/free-solid-svg-icons'
 
 // prettier-ignore
@@ -33,7 +33,7 @@ library.add(
     faPlayCircle, faStopCircle, faPause, faDatabase, faSave, faCamera, faHistory, faSkullCrossbones, faUndoAlt, 
     faSyncAlt, faPowerOff, faPencil, faArrowRight, faArrowLeft, faCompactDisc, faCheckCircle, faHdd, faMinus, faTerminal,
     faPaintbrush, faTv, faCircle, faRefresh, faCaretDown, faTimesCircle, faAngleLeft, faAngleRight, faCopy,
-    faTableColumns, faArrowPointer, faBroom, faEraser
+    faTableColumns, faArrowPointer, faBroom, faEraser, faBook
 )
 
 import App from './App.vue';